Prevent result temp file from using fds 0-2.
When opening the temp file for storing results, it can happen that
stdin/out/err is closed, and the newly-open temp file can reuse their
file descriptors.  This is mostly fine, but wreaks havoc with tests
that test the behavior of stdin/out/err.  So, take precautions to
prevent using a low fd for that temp file.

This was actually observed with the special copy of TET used for
lsb-test-core, and has been backported to tet-harness.